catalog bundle: assemble relations behind RelationBundleBuilder
All "ResolvedRelation + config -> RelationInfo" assembly now lives behind the
new final, immutable, thread-safe RelationBundleBuilder: buildRelation,
decorateColumns and its column-failure taxonomy, view/logical-schema mapping,
baseRelationInfo/attachTableStats, system-table execution resolution, and the
possession-token stamp. The UserObjectBundleService driver keeps the
knownBlobVersions slim-payload decision and the pin-identity orchestration
(scopedPinIdentity/pinIdentityFor/possessionToken/identityOnlyOrNull, chunk pin
collect/commit, flushResolutionChunk fan-out), now calling builder.build /
buildIdentityOnly / systemExecution / decorationRequired. Behavior is unchanged
(the 48-test UserObjectBundleServiceTest characterization suite stays green) and
the decoration/token logic is now covered directly by RelationBundleBuilderTest.
